dealing with 64 bit changes
new SDKs for everyone, and dealing with NSUInterger being a long
unsigned i.e. 64 bit but an int in 32 bit
http://stackoverflow.com/questions/16918826/objective-c-implicit-convers
ion-loses-integer-precision-nsuinteger-aka-unsig

+++ DD-s-Dictionary/DisplayWordViewController.m
@@ -224,7 +224,7 @@
         CGRect frame = CGRectMake(listenbutton.frame.origin.x - (buttonForLoop.frame.size.width/2 - listenbutton.frame.size.width/2), buttonForLoop.frame.origin.y, buttonForLoop.frame.size.width, buttonForLoop.frame.size.height);
         buttonForLoop.frame = frame;
     }
-    for (int i = [[self.homophonesForWord objectForKey:pronunciation] count]; i < [buttons count]; i++) {
+    for (NSUInteger i = [[self.homophonesForWord objectForKey:pronunciation] count]; i < [buttons count]; i++) {
         UIButton *buttonForLoop = [buttons objectAtIndex:i];
         buttonForLoop.hidden = YES;
     }
@@ -359,7 +359,7 @@
     
     for (NSString *pronunciation in pronunciations) {
         
-        if (([pronunciations count] > 1 && [pronunciation hasSuffix:[NSString stringWithFormat:@"%i",sender.tag]]) || ([pronunciations count] == 1)) {
+        if (([pronunciations count] > 1 && [pronunciation hasSuffix:[NSString stringWithFormat:@"%li",(long)sender.tag]]) || ([pronunciations count] == 1)) {
             [self playWord:pronunciation];
             
             //track word event with GA manual sent with Value 1
